Create Dockerfile.armv7
Adding support for armv7 to make it compatible with most of Raspbian distributions. File is exactly the same as aarch64 but taking the base image from the hypriot guys, since it supports the build of it from an x86 machine. 

I am currently running this on a RPi3 with Raspbian 9 32 bits and works perfect.