Only include client-to-client directive if exact 1
To be more consistent with the rest of the config behavior i propose this config value to be only applied if it's value is exactly 1. Otherwise it is opaque whether the config is applied or not. Providing a value of 0 to turn the config of results in turning it on.