Update tcp.md
According to documentation here (https://openvpn.net/community-resources/reference-manual-for-openvpn-2-4/), the original statement in the docs that "At this time it is not possible to generate a client config that will automatically fall back to the TCP connection" is no longer true. This can be achieved through using <connection> blocks in the oven profile. 

This change documents how to do it. It would be great to have the option to add that config creation to the scripts eventually. I've tested this config change manually and it seems to be working well.